adding functionality to make agent move to another target once reached to current target, and adding many more such agents
