--- Assets/NavMeshAgentSample.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.AI;
-
-public class NavMeshAgentSample : MonoBehaviour {
-
-    public Transform target;
-    NavMeshAgent agent;
-	// Use this for initialization
-	void Start () {
-        agent = GetComponent<NavMeshAgent>();
-	}
-	
-	// Update is called once per frame
-	void Update () {
-        agent.SetDestination(target.position);
-	}
-}
