+++ Assets/Scripts/Destinations.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ce82afc158e174c2e9c2be731b223801
+timeCreated: 1501072381
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
