+++ CS7045-Dissertation---Crowd-Simulation.sln
@@ -1,0 +1,23 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CS7045-Dissertation---Crowd-Simulation", "Assembly-CSharp.csproj", "{3C5DA656-CC18-A500-6C56-964D360A49FE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3C5DA656-CC18-A500-6C56-964D360A49FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3C5DA656-CC18-A500-6C56-964D360A49FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3C5DA656-CC18-A500-6C56-964D360A49FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3C5DA656-CC18-A500-6C56-964D360A49FE}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(MonoDevelopProperties) = preSolution
+		StartupItem = Assembly-CSharp.csproj
+	EndGlobalSection
+EndGlobal
